pyproject.toml
[project]
name = "stardew-friendship-bot"
version = "0.1.0"
description = "A domain Q&A chatbot for Stardew Valley Friendship & Gifting."
requires-python = ">=3.10"
dependencies = [
    "fastapi>=0.110.0",
    "uvicorn>=0.27.0",
    "litellm>=1.34.0",
    "python-dotenv>=1.0.1",
    "pydantic>=2.6.0",
    "pytest>=8.0.0"
]